+++ doc_layout/document_layout.py
@@ -1,4 +1,4 @@
-from layout.block_layout import BlockLayout
+from doc_layout.block_layout import BlockLayout
 
 class DocumentLayout:
     HSTEP = 13
