+++ doc_layout/draw_rect.py
