+++ doc_layout/draw_text.py
