+++ util/utils.py
@@ -97,3 +97,8 @@ def has_px_ending(property):
 
 def is_leaf(node):
     return len(node.children) == 0
+
+def paint_tree(layout_object, display_list):
+    display_list.extend(layout_object.paint())
+    for child in layout_object.children:
+        paint_tree(child, display_list)
